To get CDAKit to compile with Xcode 8.1 GM, needed to break apart record.swift into 4 pieces. It may not need to actually be in 4 pieces, as I believe the val dictionary was the culprit (its a Swift compiler bug in 8.1) but now it cleanly compiles and all tests succeed in Xcode 8.1 GM
